Rewrite LineSwitchService / FailureFilter
Fixes: #72
