+++ packages/wotan/index.ts
@@ -12,13 +12,12 @@ export * from './src/services/default/message-handler';
 export * from './src/services/default/resolver';
 export * from './src/services/default/rule-loader-host';
 export * from './src/services/default/configuration-provider';
-export * from './src/services/default/line-switch-parser';
+export * from './src/services/default/line-switches';
 export * from './src/services/cached-file-system';
 export * from './src/services/configuration-manager';
 export * from './src/services/formatter-loader';
 export * from './src/services/processor-loader';
 export * from './src/services/rule-loader';
-export * from './src/services/line-switches';
 export * from './src/linter';
 export * from './src/runner';
 export * from './src/test';
